% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital_vol1.R
\docType{data}
\name{capital_vol1}
\alias{capital_vol1}
\title{Capital, Volume I, by Karl Marx (1867)}
\format{
A named list, containing:
\describe{
\item{toc}{Table of Contents: a character vector}
\item{body}{Main text, named list, separated by chapter}
\item{footnotes}{Footnotes, named list, separated by chapter}
\item{credits}{credits, a character vector of length 1}
}
}
\source{
\url{https://www.marxists.org/archive/marx/works/1867-c1/}
}
\usage{
capital_vol1
}
\description{
This version of Capital was taken from the Marx/Engels Internet Archive
(marxists.org), which was based on the first English edition (1887),
translated by Samuel Moore and Edward Aveling, and edited by Frederick
Engels.
}
\details{
To facilitate analysis and locating of footnotes (don't skip the footnotes
on a read!), the main text and footnotes are provided as named lists, with
the names taking the form "Chapter_01" to "Chapter_33".
}
\keyword{datasets}
