% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital.R
\name{capital_vol1_book}
\alias{capital_vol1_book}
\title{Combine Capital Vol I components into a single character vector}
\usage{
capital_vol1_book()
}
\description{
The data provided in this package separates out the components of the
text into lists to facilitate analysis, but it can also be handy to have
all of it in a single character vector. For example, you can then use a
function like \code{writeLines} to print it as a text file!
}
