% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital.R
\docType{data}
\name{capital_vol2}
\alias{capital_vol2}
\title{Capital, Volume II (1885)}
\format{
A data frame, containing four columns:
\describe{
\item{section}{
Section of book, one of:
"toc" (Table of Contents),
"body" (main text),
"footnotes" (footnotes),
"credits" (list of transcribers and HTML markup contributors)
}
\item{part}{Part of text; the main text is divided into 8 parts}
\item{chapter}{Chapter of text: the main text is divided into 33 chapters}
\item{text}{Text, usually about one paragraph or footnote per row. (This
depends on how the HTML markup was done.)}
}
}
\source{
\url{https://www.marxists.org/archive/marx/works/1867-c1/}
}
\usage{
capital_vol2
}
\description{
This volume of Capital was written by Marx in 1863-1878, and edited by Engels
for publication in 1885. This version of Capital was taken from the
Marx/Engels Internet Archive (marxists.org), and is based on the revised
second edition (1893).
}
\details{
The text is stored as a data frame with four columns. The main text is
included in the "body" section. Note that the footnote numbering re-starts at
the beginning of each chapter.
}
\keyword{datasets}
