% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxmywords-pkg.R
\docType{package}
\name{marxmywords}
\alias{marxmywords}
\title{Marx My Words}
\description{
This package contains the complete text of Karl Marx' Capital Volume I,
formatted to be convenient for text analysis.
}
