% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital_vol1.R
\docType{data}
\name{capital_vol1}
\alias{capital_vol1}
\title{Capital, Volume I, by Karl Marx (1867)}
\format{
A data frame, containing four columns:
\describe{
\item{section}{
Section of book, one of:
"toc" (Table of Contents),
"body" (main text),
"footnotes" (footnotes),
"credits" (list of transcribers and HTML markup contributors)
}
\item{part}{Part of text; the main text is divided into 8 parts}
\item{chapter}{Chapter of text: the main text is divided into 33 chapters}
\item{text}{Text, usually about one paragraph or footnote per row. (This
depends on how the HTML markup was done.)}
}
}
\source{
\url{https://www.marxists.org/archive/marx/works/1867-c1/}
}
\usage{
capital_vol1
}
\description{
This version of Capital was taken from the Marx/Engels Internet Archive
(marxists.org), which was based on the first English edition (1887),
translated by Samuel Moore and Edward Aveling, and edited by Frederick
Engels.
}
\details{
The text is stored as a data frame with four columns. The main text is
included in the "body" section, but don't skip the footnotes on a read,
they're a lot of fun. Note that the footnote numbering re-starts at the
beginning of each chapter.
}
\keyword{datasets}
